#231 | Fixed typos and removed unnecessary messages
